Move to using beets to disambiguate album
Most notoriously Weezer have a lot of albums with the same name. Two of
them were even released in the same year, so the year would not be
enough to tell them apart. As far as I could tell there is no field
available through MPD that would fix this.

Unfortunately beets lookups are very slow, so there is a delay both
spawning the menu and especially adding the album. This could be solved
by keeping our own copy of the beets database (probably best kept
up-to-date by a beets plugin). Ideally it would be solved in beets, but
there seems to have been no progress on this in many years
(https://github.com/beetbox/beets/issues/2388).